hw_back_14/pyproject.toml
[tool.poetry]
name = "hw-back-14"
version = "0.1.0"
description = ""
authors = ["Zhangeldi <zhanikshakimurat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
